+++ apps/website/docusaurus.config.ts
@@ -16,11 +16,13 @@ const config: Config = {
     locales: ["en"],
   },
   markdown: {
+    hooks: {
+      onBrokenMarkdownLinks: "warn",
+    },
     mermaid: true,
   },
   onBrokenAnchors: "throw",
   onBrokenLinks: "throw",
-  onBrokenMarkdownLinks: "warn",
 
   // GitHub pages deployment config.
   // If you aren't using GitHub pages, you don't need these.
@@ -93,13 +95,9 @@ const config: Config = {
       "classic",
       {
         blog: {
-          // Remove this to remove the "edit this page" links.
-          editUrl: "https://github.com/pagopa/dx/tree/main/website/",
           showReadingTime: true,
         },
         docs: {
-          // Remove this to remove the "edit this page" links.
-          editUrl: "https://github.com/pagopa/dx/tree/main/website/",
           sidebarCollapsed: true,
           sidebarCollapsible: true,
           sidebarPath: "./sidebars.ts",