+++ infra/modules/azure_function_app/CHANGELOG.md
@@ -1,5 +1,17 @@
 # azure_function_app
 
+## 4.1.1
+
+### Patch Changes
+
+- bdf1e20: Fix private_dns_zone_ids validation for older Terraform versions
+
+## 4.1.0
+
+### Minor Changes
+
+- ce7ad91: Add variables to support private DNS zones in different subscriptions
+
 ## 4.0.0
 
 ### Major Changes